+++ src/main/java/com/johnsproject/jpge2/dto/Model.java
@@ -31,8 +31,8 @@ public class Model extends SceneObject {
 	private Vertex[] vertices;
 	private Face[] faces;
 	private Material[] materials;
-	private long[][] modelMatrix = MatrixProcessor.generate();
-	private long[][] normalMatrix = MatrixProcessor.generate();
+	private int[][] modelMatrix = MatrixProcessor.generate();
+	private int[][] normalMatrix = MatrixProcessor.generate();
 	
 	public Model (String name, Transform transform, Vertex[] vertices, Face[] faces, Material[] materials) {
 		super(name, transform);
@@ -71,14 +71,14 @@ public class Model extends SceneObject {
 		return materials[index];
 	}
 
-	public long[][] getModelMatrix() {
+	public int[][] getModelMatrix() {
 		if (this.hasChanged()) {
 			GraphicsProcessor.worldMatrix(modelMatrix, this);
 		}
 		return modelMatrix;
 	}
 	
-	public long[][] getNormalMatrix() {
+	public int[][] getNormalMatrix() {
 		if (this.hasChanged()) {
 			GraphicsProcessor.normalMatrix(normalMatrix, this);
 		}
