+++ src/main/java/com/johnsproject/jpge2/importers/SOMImporter.java
@@ -47,14 +47,14 @@ public class SOMImporter {
 		String[] vNormalData = rawData.split("vNormal<")[1].split(">vNormal", 2)[0].split(",");
 		String[] vMaterialData = rawData.split("vMaterial<")[1].split(">vMaterial", 2)[0].split(",");
 		for (int i = 0; i < vertices.length * 3; i += 3) {
-			long[] location = VectorProcessor.generate();
-			location[vx] = (long)(getFloat(vLocationData[i + vx]) * MathProcessor.FP_VALUE);
-			location[vy] = (long)(getFloat(vLocationData[i + vy]) * MathProcessor.FP_VALUE);
-			location[vz] = (long)(getFloat(vLocationData[i + vz]) * MathProcessor.FP_VALUE);
-			long[] normal = VectorProcessor.generate();
-			normal[vx] = (long)(getFloat(vNormalData[i + vx]) * MathProcessor.FP_VALUE);
-			normal[vy] = (long)(getFloat(vNormalData[i + vy]) * MathProcessor.FP_VALUE);
-			normal[vz] = (long)(getFloat(vNormalData[i + vz]) * MathProcessor.FP_VALUE);
+			int[] location = VectorProcessor.generate();
+			location[vx] = (int)(getFloat(vLocationData[i + vx]) * MathProcessor.FP_VALUE);
+			location[vy] = (int)(getFloat(vLocationData[i + vy]) * MathProcessor.FP_VALUE);
+			location[vz] = (int)(getFloat(vLocationData[i + vz]) * MathProcessor.FP_VALUE);
+			int[] normal = VectorProcessor.generate();
+			normal[vx] = (int)(getFloat(vNormalData[i + vx]) * MathProcessor.FP_VALUE);
+			normal[vy] = (int)(getFloat(vNormalData[i + vy]) * MathProcessor.FP_VALUE);
+			normal[vz] = (int)(getFloat(vNormalData[i + vz]) * MathProcessor.FP_VALUE);
 			int material = getint(vMaterialData[i / 3]);
 			vertices[i / 3] = new Vertex(i / 3, location, normal, material);
 		}
@@ -77,19 +77,19 @@ public class SOMImporter {
 			int vertex2 = getint(fVertex2Data[i / 6]);
 			int vertex3 = getint(fVertex3Data[i / 6]);
 			int material = getint(fMaterialData[i / 6]);
-			long[] normal = VectorProcessor.generate();
-			normal[vx] = (long)(getFloat(fNormalData[(i / 2) + vx]) * MathProcessor.FP_VALUE);
-			normal[vy] = (long)(getFloat(fNormalData[(i / 2) + vy]) * MathProcessor.FP_VALUE);
-			normal[vz] = (long)(getFloat(fNormalData[(i / 2) + vz]) * MathProcessor.FP_VALUE);
-			long[] uv1 = VectorProcessor.generate();
-			uv1[vx] = (long)(getFloat(fUV1Data[(i / 3) + vx]) * MathProcessor.FP_VALUE);
-			uv1[vy] = (long)(getFloat(fUV1Data[(i / 3) + vy]) * MathProcessor.FP_VALUE);
-			long[] uv2 = VectorProcessor.generate();
-			uv2[vx] = (long)(getFloat(fUV2Data[(i / 3) + vx]) * MathProcessor.FP_VALUE);
-			uv2[vy] = (long)(getFloat(fUV2Data[(i / 3) + vy]) * MathProcessor.FP_VALUE);
-			long[] uv3 = VectorProcessor.generate();
-			uv3[vx] = (long)(getFloat(fUV3Data[(i / 3) + vx]) * MathProcessor.FP_VALUE);
-			uv3[vy] = (long)(getFloat(fUV3Data[(i / 3) + vy]) * MathProcessor.FP_VALUE);
+			int[] normal = VectorProcessor.generate();
+			normal[vx] = (int)(getFloat(fNormalData[(i / 2) + vx]) * MathProcessor.FP_VALUE);
+			normal[vy] = (int)(getFloat(fNormalData[(i / 2) + vy]) * MathProcessor.FP_VALUE);
+			normal[vz] = (int)(getFloat(fNormalData[(i / 2) + vz]) * MathProcessor.FP_VALUE);
+			int[] uv1 = VectorProcessor.generate();
+			uv1[vx] = (int)(getFloat(fUV1Data[(i / 3) + vx]) * MathProcessor.FP_VALUE);
+			uv1[vy] = (int)(getFloat(fUV1Data[(i / 3) + vy]) * MathProcessor.FP_VALUE);
+			int[] uv2 = VectorProcessor.generate();
+			uv2[vx] = (int)(getFloat(fUV2Data[(i / 3) + vx]) * MathProcessor.FP_VALUE);
+			uv2[vy] = (int)(getFloat(fUV2Data[(i / 3) + vy]) * MathProcessor.FP_VALUE);
+			int[] uv3 = VectorProcessor.generate();
+			uv3[vx] = (int)(getFloat(fUV3Data[(i / 3) + vx]) * MathProcessor.FP_VALUE);
+			uv3[vy] = (int)(getFloat(fUV3Data[(i / 3) + vy]) * MathProcessor.FP_VALUE);
 			faces[i / 6] = new Face(i / 6, vertex1, vertex2, vertex3, material, normal, uv1, uv2, uv3);
 		}
 		return faces;
@@ -102,7 +102,7 @@ public class SOMImporter {
 		String[] mDiffuseIntensityData = rawData.split("mDiffuseIntensity<")[1].split(">mDiffuseIntensity", 2)[0].split(",");
 		String[] mSpecularIntensityData = rawData.split("mSpecularIntensity<")[1].split(">mSpecularIntensity", 2)[0].split(",");
 		for (int i = 0; i < materials.length * 4; i+=4) {
-			// *255 to get long rgb values
+			// *255 to get int rgb values
 			int r = (int)(getFloat(mDiffuseColorData[i]) * 256);
 			int	g = (int)(getFloat(mDiffuseColorData[i+1]) * 256);
 			int	b = (int)(getFloat(mDiffuseColorData[i+2]) * 256);
